Save new users a lot of pain from rug pullers
At onboarding, users are excited to get started, but the first thing they HAVE to do is update their profile. 
The initial FR should not be anything other than 100% at this point in the process!

If changed from 100% at this first opportunity, scammers using auto-bots quickly purchase before users are able to buy their coin.
Before they know it, upon trying to take their 2nd action, their coin price has increased and any $ they put in goes directly into the hands of the owners of the bots.

This 1 change is a must imho to improve the user experience of new clouters.